update image ratio to float

+++ features/user/notifcenter/src/main/java/com/tokopedia/notifcenter/data/entity/notification/Ratio.kt
@@ -5,7 +5,7 @@ import com.google.gson.annotations.SerializedName
 
 data class Ratio(
         @SerializedName("x")
-        val x: Int = 2,
+        val x: Float = 2f,
         @SerializedName("y")
-        val y: Int = 1
+        val y: Float = 1f
 )
